% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{calls}
\alias{calls}
\alias{sn_tickers}
\alias{sn_category}
\title{Call}
\usage{
sn_tickers(q, ..., pages = 1)

sn_category(q = "alltickers", ..., pages = 1)
}
\arguments{
\item{q}{Vector of tickers or a valid section (\code{alltickers}, \code{general}).}

\item{...}{Any other parameter from \url{https://stocknewsapi.com/documentation}.}

\item{pages}{Number of pages to retrieve.}
}
\description{
Get news by tickers and categories.
}
\examples{
\dontrun{
tickers <- sn_tickers(c("AMZN", "GOOG"))
sections <- sn_category("general")  
}

}
