% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{sn_tickers}
\alias{sn_tickers}
\title{Tickers}
\usage{
sn_tickers(tickers, ..., pages = 1)
}
\arguments{
\item{tickers}{Vector of tickers.}

\item{...}{Any other parameter from \url{https://stocknewsapi.com/documentation}.}

\item{pages}{Number of pages to retrieve.}
}
\description{
Get tickers data.
}
